Refactor LambdaEventHandler to be inside LambdaExecutor; Remove bytebuf inside LambdaEvent.
