+++ runtime/master/pom.xml
@@ -97,5 +97,11 @@ under the License.
       <artifactId>aws-java-sdk-lambda</artifactId>
       <version>1.11.429</version>
     </dependency>
+    <dependency>
+      <groupId>log4j</groupId>
+      <artifactId>log4j</artifactId>
+      <version>1.2.17</version>
+      <scope>compile</scope>
+    </dependency>
   </dependencies>
 </project>
